Create Multi_Sensors_test.ino
This code was used in our initial testing of two sensors chained together and providing an output from both sensors.  The goal of the project is to have the two head sensors work together to establish where an obstacle/object is located.  The hip sensor will detect drop offs with the assistance of an algorithm.  The algorithm will need to retain the last output value and evaluate if the change in the next output value (a sudden increase in distance) indicates a drop off.