New Code Files
After some review of sample code previously it was determined that to combat the issues developed from testing we would need to develop code that would conquer all of our issues.  Instead of altering existing code we were able to develop this base level code to function as we desired.  From this point on code will be updated from these files.