New Code Files
These files include the alert system we have created for the sensors to notify the user.  This includes determining the order in which alerts will happen, the hip sensor also includes the angled calculations that need to occur upon start-up.
It has been determined that we will need to run 2 arduino's in our system so we have designed an initial master and slave arduino code for two arduino's to communicate with one another.  One arduino will run the wave shield and will be the slave arduino that will receive inputs from the master arduino, which will operate the sensors.  The inputs will designate which audio file to play based on the readings from the sensors.